requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.1
gunicorn==23.0.0
lxml==5.3.0
numpy==2.1.1
packaging==24.2
pandas==2.2.2
pillow==10.4.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
reportlab==4.2.5
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
